requirements.txt
pygame
wtforms
Flask-WTF
beautifulsoup4
email_validator
uvicorn
fastapi